next

+++ src/description.rs
@@ -10,14 +10,14 @@ pub type NodeID = String;
 pub struct Node {
     pub x: f64,
     pub y: f64,
-    pub demand: isize,
+    pub demand: i32,
     pub is_depot: bool,
 }
 
 #[derive(Serialize, Deserialize, Debug)]
 #[serde(rename_all = "camelCase")]
 pub struct Description {
-    pub capacity: isize,
+    pub capacity: i32,
     pub edge_weight_type: String,
     pub nodes: HashMap<NodeID, Node>,
 }
